# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Biblioteca do Condomínio API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PaginationMeta {
  page: number;
  limit: number;
  total: number;
  totalPages: number;
}

export type BookStatus = (typeof BookStatus)[keyof typeof BookStatus];

export const BookStatus = {
  draft: "draft",
  available: "available",
  borrowed: "borrowed",
  reserved: "reserved",
  lost: "lost",
  unavailable: "unavailable",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  pending: "pending",
  active: "active",
  inactive: "inactive",
  blocked: "blocked",
} as const;

export type LoanStatus = (typeof LoanStatus)[keyof typeof LoanStatus];

export const LoanStatus = {
  active: "active",
  returned: "returned",
  overdue: "overdue",
} as const;

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  waiting: "waiting",
  notified: "notified",
  fulfilled: "fulfilled",
  cancelled: "cancelled",
} as const;

export type NoteType = (typeof NoteType)[keyof typeof NoteType];

export const NoteType = {
  rule: "rule",
  info: "info",
  announcement: "announcement",
} as const;

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AdminProfile {
  id: number;
  name: string;
  email: string;
  createdAt: string;
  updatedAt: string;
}

export interface AdminLoginResponse {
  token: string;
  admin: AdminProfile;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  block?: string | null;
  house?: string | null;
  status: UserStatus;
  createdAt: string;
  updatedAt: string;
}

export interface ReaderLoginResponse {
  token: string;
  user: UserProfile;
}

export interface AdminsListResponse {
  admins: AdminProfile[];
}

export interface CreateAdminRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface BookSummary {
  id: number;
  title: string;
  author: string;
  genre: string;
  isbn?: string | null;
  publishedYear?: number | null;
  imageUrl?: string | null;
  status: BookStatus;
}

export interface LoanSummary {
  id: number;
  bookId: number;
  userId: number;
  status: LoanStatus;
  loanDate: string;
  dueDate: string;
  returnDate?: string | null;
  book?: BookSummary;
  user?: UserProfile;
}

export type UserDetail = UserProfile & {
  loans: LoanSummary[];
};

export interface UsersListResponse {
  users: UserProfile[];
  meta: PaginationMeta;
}

export interface CreateUserRequest {
  name: string;
  email: string;
  phone?: string | null;
  block?: string | null;
  house?: string | null;
  status?: UserStatus;
  password?: string | null;
}

export interface UpdateUserRequest {
  name?: string;
  email?: string;
  phone?: string | null;
  block?: string | null;
  house?: string | null;
  status?: UserStatus;
  password?: string | null;
}

export interface ReservationDetail {
  id: number;
  bookId: number;
  userId: number;
  position: number;
  status: ReservationStatus;
  notifiedAt?: string | null;
  expiresAt?: string | null;
  createdAt: string;
  updatedAt: string;
  book?: BookSummary;
  user?: UserProfile;
}

export type BookDetail = BookSummary & {
  description?: string | null;
  createdAt: string;
  updatedAt: string;
  loans: LoanSummary[];
  reservations: ReservationDetail[];
};

export interface BooksListResponse {
  books: BookSummary[];
  meta: PaginationMeta;
}

export interface CreateBookRequest {
  title: string;
  author: string;
  genre: string;
  isbn?: string | null;
  publishedYear?: number | null;
  imageUrl?: string | null;
  description?: string | null;
  status?: BookStatus;
}

export interface UpdateBookRequest {
  title?: string;
  author?: string;
  genre?: string;
  isbn?: string | null;
  publishedYear?: number | null;
  imageUrl?: string | null;
  description?: string | null;
  status?: BookStatus;
}

export type LoanDetail = LoanSummary & {
  createdAt: string;
};

export interface LoansListResponse {
  loans: LoanDetail[];
  meta: PaginationMeta;
}

export interface CreateLoanRequest {
  bookId: number;
  userId: number;
  dueDate?: string;
}

export interface ReservationsListResponse {
  reservations: ReservationDetail[];
}

export interface CreateReservationRequest {
  bookId: number;
  userId: number;
}

export type ReservationActionRequestAction =
  (typeof ReservationActionRequestAction)[keyof typeof ReservationActionRequestAction];

export const ReservationActionRequestAction = {
  notify: "notify",
  cancel: "cancel",
  fulfill: "fulfill",
  advance: "advance",
} as const;

export interface ReservationActionRequest {
  action: ReservationActionRequestAction;
}

export interface ReaderDashboard {
  activeLoans: number;
  overdueLoans: number;
  pendingReservations: number;
  totalLoans: number;
}

export interface ReaderLoansResponse {
  active: LoanDetail[];
  history: LoanDetail[];
}

export interface ReaderReservationsResponse {
  reservations: ReservationDetail[];
}

export interface ReaderCreateReservationRequest {
  bookId: number;
}

export interface ReaderUpdateProfileRequest {
  phone?: string | null;
  password?: string | null;
  currentPassword?: string | null;
}

export interface ReaderLookupRequest {
  email: string;
}

export interface ReaderLookupResponse {
  user: UserProfile;
  activeLoans: LoanDetail[];
  overdueLoans: LoanDetail[];
  reservations: ReservationDetail[];
}

export interface LibraryDashboard {
  totalBooks: number;
  availableBooks: number;
  borrowedBooks: number;
  activeReaders: number;
  activeLoans: number;
  overdueLoans: number;
  popularBooks: BookSummary[];
}

export interface Note {
  id: number;
  title: string;
  content: string;
  type: NoteType;
  active: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface NotesListResponse {
  notes: Note[];
}

export interface CreateNoteRequest {
  title: string;
  content: string;
  type?: NoteType;
  active?: boolean;
}

export interface UpdateNoteRequest {
  title?: string;
  content?: string;
  type?: NoteType;
  active?: boolean;
}

export interface GeminiBookSearchRequest {
  query: string;
}

export interface GeminiBookSearchResponse {
  title?: string | null;
  author?: string | null;
  genre?: string | null;
  publishedYear?: number | null;
  isbn?: string | null;
  description?: string | null;
  imageUrl?: string | null;
}

export type ListBooksParams = {
  search?: string;
  status?: BookStatus;
  genre?: string;
  page?: number;
  limit?: number;
};

export type ListUsersParams = {
  search?: string;
  status?: UserStatus;
  page?: number;
  limit?: number;
};

export type ListLoansParams = {
  status?: LoanStatus;
  userId?: number;
  bookId?: number;
  dueSoon?: boolean;
  dueDateFrom?: string;
  dueDateTo?: string;
  reportStatus?: ListLoansReportStatus;
  page?: number;
  limit?: number;
};

export type ListLoansReportStatus =
  (typeof ListLoansReportStatus)[keyof typeof ListLoansReportStatus];

export const ListLoansReportStatus = {
  upcoming: "upcoming",
  due_today: "due_today",
  overdue: "overdue",
  returned: "returned",
} as const;

export type ListReservationsParams = {
  bookId?: number;
  userId?: number;
  status?: ReservationStatus;
};

export type ListNotesParams = {
  /**
   * Include inactive notes (admin use)
   */
  all?: boolean;
};