# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Biblioteca do Condomínio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: adminMe
      tags: [auth]
      summary: Get current admin
      security:
        - adminBearerAuth: []
      responses:
        "200":
          description: Current admin data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Books ───────────────────────────────────────────────────────────────────
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List books
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/BookStatus"
        - name: genre
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Book list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BooksListResponse"
    post:
      operationId: createBook
      tags: [books]
      summary: Create book
      security:
        - adminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookRequest"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get book detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book detail with loan history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update book
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookRequest"
      responses:
        "200":
          description: Book updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete book
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "409":
          description: Cannot delete (has loan history)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Users ───────────────────────────────────────────────────────────────────
...
  /admins:
    get:
      operationId: listAdmins
      tags: [admins]
      summary: List admins
      security:
        - adminBearerAuth: []
      responses:
        "200":
          description: Admin list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminsListResponse"
    post:
      operationId: createAdmin
      tags: [admins]
      summary: Create admin
      security:
        - adminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminRequest"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admins/{id}:
    delete:
      operationId: deleteAdmin
      tags: [admins]
      summary: Delete admin
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Cannot delete self
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Gemini ───────────────────────────────────────────────────────────────────
  /gemini/book-search:
    post:
      operationId: geminiBookSearch
      tags: [gemini]
      summary: AI book data lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiBookSearchRequest"
      responses:
        "200":
          description: Book data from AI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiBookSearchResponse"
        "500":
          description: AI error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    readerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ─── Enums ─────────────────────────────────────────────────────────────────
    BookStatus:
      type: string
      enum: [draft, available, borrowed, reserved, lost, unavailable]
    UserStatus:
      type: string
      enum: [pending, active, inactive, blocked]
    LoanStatus:
      type: string
      enum: [active, returned, overdue]
    ReservationStatus:
      type: string
      enum: [waiting, notified, fulfilled, cancelled]
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/AdminProfile"
      required: [token, admin]
...
    AdminProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, createdAt, updatedAt]
    AdminsListResponse:
      type: object
      properties:
        admins:
          type: array
          items:
            $ref: "#/components/schemas/AdminProfile"
      required: [admins]
    CreateAdminRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [name, email, password]

    # ─── Users ─────────────────────────────────────────────────────────────────
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        block:
          type: string
          nullable: true
        house:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/UserStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, status, createdAt, updatedAt]
...
    BookSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        isbn:
          type: string
          nullable: true
        publishedYear:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/BookStatus"
      required: [id, title, author, genre, status]
    BookDetail:
      allOf:
        - $ref: "#/components/schemas/BookSummary"
        - type: object
          properties:
            description:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            loans:
              type: array
              items:
                $ref: "#/components/schemas/LoanSummary"
            reservations:
              type: array
              items:
                $ref: "#/components/schemas/ReservationDetail"
          required: [createdAt, updatedAt, loans, reservations]
    BooksListResponse:
      type: object
      properties:
        books:
          type: array
          items:
            $ref: "#/components/schemas/BookSummary"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [books, meta]
    CreateBookRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        isbn:
          type: string
          nullable: true
        publishedYear:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/BookStatus"
      required: [title, author, genre]
    UpdateBookRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        isbn:
          type: string
          nullable: true
        publishedYear:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/BookStatus"

    # ─── Loans ─────────────────────────────────────────────────────────────────
    LoanSummary:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        status:
          $ref: "#/components/schemas/LoanStatus"
        loanDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
          nullable: true
        book:
          $ref: "#/components/schemas/BookSummary"
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [id, bookId, userId, status, loanDate, dueDate]
...
    ReservationDetail:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        position:
          type: integer
        status:
          $ref: "#/components/schemas/ReservationStatus"
        notifiedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        book:
          $ref: "#/components/schemas/BookSummary"
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [id, bookId, userId, position, status, createdAt, updatedAt]
...
    GeminiBookSearchRequest:
      type: object
      properties:
        query:
          type: string
      required: [query]
    GeminiBookSearchResponse:
      type: object
      properties:
        title:
          type: string
          nullable: true
        author:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        publishedYear:
          type: integer
          nullable: true
        isbn:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required: []