# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Biblioteca do Condomínio API
 * OpenAPI spec version: 1.0.0
 */

export interface LoginRequest {
  email: string;
  password: string;
}