# lib/api-zod/src/generated/types/reservationDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Biblioteca do Condomínio API
 * OpenAPI spec version: 1.0.0
 */
import type { BookSummary } from "./bookSummary";
import type { ReservationStatus } from "./reservationStatus";
import type { UserProfile } from "./userProfile";

export interface ReservationDetail {
  id: number;
  bookId: number;
  userId: number;
  position: number;
  status: ReservationStatus;
  notifiedAt?: Date | null;
  expiresAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
  book?: BookSummary;
  user?: UserProfile;
}