# lib/api-zod/src/generated/types/reservationStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Biblioteca do Condomínio API
 * OpenAPI spec version: 1.0.0
 */

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  waiting: "waiting",
  notified: "notified",
  fulfilled: "fulfilled",
  cancelled: "cancelled",
} as const;